requirements.txt
streamlit
langchain
langchain-community
langchain-openai
langchain-google-genai
faiss-cpu
pypdf
python-dotenv
tiktoken
openai
